+++ tests/decorators/Emit.spec.ts
@@ -1,6 +1,6 @@
 import { mount, Wrapper } from '@vue/test-utils'
 import Vue, { CreateElement } from 'vue'
-import Component from 'vue-class-component'
+import Component from '@havunen/vue2-class-component'
 import { Emit } from '../../src/decorators/Emit'
 
 const mockFn = jest.fn()