+++ tests/decorators/Ref.spec.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import Component from 'vue-class-component'
+import Component from '@havunen/vue2-class-component'
 import { Ref } from '../../src/decorators/Ref'
 
 describe(Ref, () => {